requirements.txt
anthropic>=0.40.0
pywin32>=306
python-dotenv>=1.0.0
